Move DeleteWarningModal to Components
