+++ client/components/DeleteWarningModal.js
@@ -3,7 +3,7 @@ import React from 'react';
 
 import { useTranslation } from '../contexts/TranslationContext';
 
-const DeleteWarningModal = ({ onDelete, onCancel, ...props }) => {
+const DeleteWarningModal = ({ onDelete, onCancel, children, ...props }) => {
 	const t = useTranslation();
 	return <Modal {...props}>
 		<Modal.Header>
@@ -11,6 +11,9 @@ const DeleteWarningModal = ({ onDelete, onCancel, ...props }) => {
 			<Modal.Title>{t('Are_you_sure')}</Modal.Title>
 			<Modal.Close onClick={onCancel}/>
 		</Modal.Header>
+		<Modal.Content>
+			{children}
+		</Modal.Content>
 		<Modal.Footer>
 			<ButtonGroup align='end'>
 				<Button ghost onClick={onCancel}>{t('Cancel')}</Button>
